+++ tests/features/test_apis.py
@@ -1,4 +1,5 @@
 from typing import Annotated
+import pytest
 
 import ididi
 from ididi import Graph, use
@@ -16,5 +17,6 @@ def main(user_service: Annotated[UserService, use(UserService)]):
     return user_service
 
 
+@pytest.mark.debug
 def test_ididi_entry():
     assert isinstance(ididi.entry(main)(), UserService)